style(AutoComplete): extract types and helpers, move into separate dir
