+++ packages/react-components/src/components/Input/AutoComplete/index.ts
@@ -1,0 +1,5 @@
+import { AutoComplete } from './AutoComplete';
+import { AutoCompleteProps } from './types';
+
+export { AutoComplete };
+export type { AutoCompleteProps };
